+++ sequence/sequences/integer/explicit.py
@@ -2,13 +2,18 @@ from typing import Any
 
 from sequence.core.mixin import MonotonicIncreasingMixin
 from sequence.core.infinite_type import Explicit
-from sequence.sequences.integer.explicit_generalised import PolygonalNumbers, GeneralisedNexusNumbers
+from sequence.core.utils.functions import digit_sum
+from sequence.sequences.integer.explicit_generalised import (
+    DigitSumSequence,
+    PolygonalNumbers,
+    GeneralisedNexusNumbers,
+)
 
 
 class A000027(Explicit):
     """The natural numbers (https://oeis.org/A000027)."""
 
-    sequence_name = "natural numbers"
+    SEQUENCE_NAME = "natural numbers"
 
     def __contains__(self, item: Any) -> bool:
         return True
@@ -21,10 +26,19 @@ PositiveIntegers = A000027
 NaturalNumbers = A000027
 
 
+class A000120(DigitSumSequence):
+    """Number of 1's in binary expansion of n (https://oeis.org/A000120)."""
+
+    SEQUENCE_NAME = "sequence A000120"
+
+    def __init__(self) -> None:
+        super().__init__(base=2)
+
+
 class A000217(PolygonalNumbers):
     """Triangular numbers (https://oeis.org/A000217)."""
 
-    sequence_name = "triangular numbers"
+    SEQUENCE_NAME = "triangular numbers"
 
     def __init__(self) -> None:
         super().__init__(number_of_sides=3)
@@ -36,7 +50,7 @@ TriangularNumbers = A000217
 class A000290(Explicit):
     """Square numbers (https://oeis.org/A000290)."""
 
-    sequence_name = "square numbers"
+    SEQUENCE_NAME = "square numbers"
 
     def __contains__(self, item: Any) -> bool:
         return False if item < 0 else int(item ** (1 / 2)) == item ** (1 / 2)
@@ -51,7 +65,7 @@ SquareNumbers = A000290
 class A000326(PolygonalNumbers):
     """Pentagonal numbers (https://oeis.org/A000326)."""
 
-    sequence_name = "pentagonal numbers"
+    SEQUENCE_NAME = "pentagonal numbers"
 
     def __init__(self) -> None:
         super().__init__(number_of_sides=5)
@@ -63,7 +77,7 @@ PentagonalNumbers = A000326
 class A000384(PolygonalNumbers):
     """Hexagonal numbers (https://oeis.org/A000384)."""
 
-    sequence_name = "hexagonal numbers"
+    SEQUENCE_NAME = "hexagonal numbers"
 
     def __init__(self) -> None:
         super().__init__(number_of_sides=6)
@@ -75,7 +89,7 @@ HexagonalNumbers = A000384
 class A000566(PolygonalNumbers):
     """Heptagonal numbers (https://oeis.org/A000566)"""
 
-    sequence_name = "heptagonal numbers"
+    SEQUENCE_NAME = "heptagonal numbers"
 
     def __init__(self) -> None:
         super().__init__(number_of_sides=7)
@@ -87,7 +101,7 @@ HeptagonalNumbers = A000566
 class A000567(PolygonalNumbers):
     """Octagonal numbers (https://oeis.org/A000567)"""
 
-    sequence_name = "octagonal numbers"
+    SEQUENCE_NAME = "octagonal numbers"
 
     def __init__(self) -> None:
         super().__init__(number_of_sides=8)
@@ -99,7 +113,7 @@ OctagonalNumbers = A000567
 class A001045(MonotonicIncreasingMixin, Explicit):
     """Jacobsthal numbers (https://oeis.org/A001045)."""
 
-    sequence_name = "Jacobsthal numbers"
+    SEQUENCE_NAME = "Jacobsthal numbers"
 
     def formula(self, index: int) -> int:
         return round(2**index / 3)
@@ -112,7 +126,7 @@ JacobsthalSequence = A001045
 class A001106(PolygonalNumbers):
     """Nonagonal numbers (https://oeis.org/A001106)"""
 
-    sequence_name = "nonagonal numbers"
+    SEQUENCE_NAME = "nonagonal numbers"
 
     def __init__(self) -> None:
         super().__init__(number_of_sides=9)
@@ -124,7 +138,7 @@ NonagonalNumbers = A001106
 class A001107(PolygonalNumbers):
     """Decagonal numbers (https://oeis.org/A001107)"""
 
-    sequence_name = "decagonal numbers"
+    SEQUENCE_NAME = "decagonal numbers"
 
     def __init__(self) -> None:
         super().__init__(number_of_sides=10)
@@ -136,7 +150,7 @@ DecagonalNumbers = A001107
 class A003215(GeneralisedNexusNumbers):
     """Hex (or centered hexagonal) numbers (https://oeis.org/A003215)."""
 
-    sequence_name = "hex numbers"
+    SEQUENCE_NAME = "hex numbers"
 
     def __init__(self) -> None:
         super().__init__(dimension=2)
@@ -155,7 +169,7 @@ CenteredHexagonalNumbers = A003215
 class A005408(Explicit):
     """The odd numbers (https://oeis.org/A005408)."""
 
-    sequence_name = "odd numbers"
+    SEQUENCE_NAME = "odd numbers"
 
     def __contains__(self, item: Any) -> bool:
         return item % 2 == 1
@@ -167,10 +181,37 @@ class A005408(Explicit):
 OddNumbers = A005408
 
 
+class A007953(DigitSumSequence):
+    """Digital sum (i.e., sum of digits) of n (https://oeis.org/A007953)."""
+
+    SEQUENCE_NAME = "digsum"
+
+    def __init__(self) -> None:
+        super().__init__(base=10)
+
+
+Digsum = A007953
+
+
+class A010060(Explicit):
+    """Thue-Morse sequence (https://oeis.org/A010060)."""
+
+    SEQUENCE_NAME = "Thue-Morse sequence"
+
+    def __contains__(self, item: Any) -> bool:
+        return item in [0, 1]
+
+    def formula(self, index: int) -> int:
+        return digit_sum(number=index, base=2) % 2
+
+
+ThueMorseSequence = A010060
+
+
 class A014551(MonotonicIncreasingMixin, Explicit):
     """Jacobsthal-Lucas numbers (https://oeis.org/A014551)."""
 
-    sequence_name = "Jacobsthal-Lucas numbers"
+    SEQUENCE_NAME = "Jacobsthal-Lucas numbers"
 
     def __contains__(self, item: Any) -> bool:
         if item == 1:
@@ -187,7 +228,7 @@ JachobsthalLucasNumbers = A014551
 class A033999(Explicit):
     """Sequence of powers of -1 (https://oeis.org/A033999)."""
 
-    sequence_name = "sequence of powers of -1"
+    SEQUENCE_NAME = "sequence of powers of -1"
 
     def __contains__(self, item: Any) -> bool:
         return item in {-1, 1}
@@ -199,7 +240,7 @@ class A033999(Explicit):
 class A051624(PolygonalNumbers):
     """Dodecagonal numbers https://oeis.org/A051624)."""
 
-    sequence_name = "dodecagonal numbers"
+    SEQUENCE_NAME = "dodecagonal numbers"
 
     def __init__(self) -> None:
         super().__init__(number_of_sides=12)
@@ -211,7 +252,7 @@ DodecagonalNumbers = A051624
 class A051682(PolygonalNumbers):
     """Hendecagonal numbers (https://oeis.org/A051682)."""
 
-    sequence_name = "hendecagonal numbers"
+    SEQUENCE_NAME = "hendecagonal numbers"
 
     def __init__(self) -> None:
         super().__init__(number_of_sides=11)
@@ -223,7 +264,7 @@ HendecagonalNumbers = A051682
 class A051865(PolygonalNumbers):
     """Tridecagonal numbers (https://oeis.org/A051865)."""
 
-    sequence_name = "tridecagonal numbers"
+    SEQUENCE_NAME = "tridecagonal numbers"
 
     def __init__(self) -> None:
         super().__init__(number_of_sides=13)
@@ -235,7 +276,7 @@ TridecagonalNumbers = A051865
 class A051866(PolygonalNumbers):
     """Tetradecagonal numbers (https://oeis.org/A051866)."""
 
-    sequence_name = "tetradecagonal numbers"
+    SEQUENCE_NAME = "tetradecagonal numbers"
 
     def __init__(self) -> None:
         super().__init__(number_of_sides=14)
@@ -247,7 +288,7 @@ TetradecagonalNumbers = A051866
 class A051867(PolygonalNumbers):
     """Pentadecagonal numbers (https://oeis.org/A051867)."""
 
-    sequence_name = "pentadecagonal numbers"
+    SEQUENCE_NAME = "pentadecagonal numbers"
 
     def __init__(self) -> None:
         super().__init__(number_of_sides=15)
@@ -259,7 +300,7 @@ PentadecagonalNumbers = A051867
 class A051868(PolygonalNumbers):
     """Hexadecagonal numbers (https://oeis.org/A051868)."""
 
-    sequence_name = "hexadecagonal numbers"
+    SEQUENCE_NAME = "hexadecagonal numbers"
 
     def __init__(self) -> None:
         super().__init__(number_of_sides=16)
@@ -271,7 +312,7 @@ HexadecagonalNumbers = A051868
 class A051869(PolygonalNumbers):
     """Heptadecagonal numbers (https://oeis.org/A051869)."""
 
-    sequence_name = "heptadecagonal numbers"
+    SEQUENCE_NAME = "heptadecagonal numbers"
 
     def __init__(self) -> None:
         super().__init__(number_of_sides=17)
@@ -283,7 +324,7 @@ HeptadecagonalNumbers = A051869
 class A051870(PolygonalNumbers):
     """Octadecagonal numbers (https://oeis.org/A051870)."""
 
-    sequence_name = "octadecagonal numbers"
+    SEQUENCE_NAME = "octadecagonal numbers"
 
     def __init__(self) -> None:
         super().__init__(number_of_sides=18)
@@ -295,7 +336,7 @@ OctadecagonalNumbers = A051870
 class A051871(PolygonalNumbers):
     """Enneadecagonal numbers (https://oeis.org/A051871)."""
 
-    sequence_name = "enneadecagonal numbers"
+    SEQUENCE_NAME = "enneadecagonal numbers"
 
     def __init__(self) -> None:
         super().__init__(number_of_sides=19)
@@ -307,7 +348,7 @@ EnneadecagonalNumbers = A051871
 class A051872(PolygonalNumbers):
     """Icosagonal numbers (https://oeis.org/A051872)."""
 
-    sequence_name = "icosagonal numbers"
+    SEQUENCE_NAME = "icosagonal numbers"
 
     def __init__(self) -> None:
         super().__init__(number_of_sides=20)
@@ -319,7 +360,7 @@ IcosagonalNumbers = A051872
 class A051873(PolygonalNumbers):
     """Icosihenagonal numbers (https://oeis.org/A051873)."""
 
-    sequence_name = "icosihenagonal numbers"
+    SEQUENCE_NAME = "icosihenagonal numbers"
 
     def __init__(self) -> None:
         super().__init__(number_of_sides=21)
@@ -331,7 +372,7 @@ IcosihenagonalNumbers = A051873
 class A051874(PolygonalNumbers):
     """Icosidigonal numbers (https://oeis.org/A051874)."""
 
-    sequence_name = "icosidigonal numbers"
+    SEQUENCE_NAME = "icosidigonal numbers"
 
     def __init__(self) -> None:
         super().__init__(number_of_sides=22)
@@ -343,7 +384,7 @@ IcosidigonalNumbers = A051874
 class A051875(PolygonalNumbers):
     """Icositrigonal numbers (https://oeis.org/A051875)."""
 
-    sequence_name = "icositrigonal numbers"
+    SEQUENCE_NAME = "icositrigonal numbers"
 
     def __init__(self) -> None:
         super().__init__(number_of_sides=23)
@@ -355,7 +396,7 @@ IcositrigonalNumbers = A051875
 class A051876(PolygonalNumbers):
     """Icositetragonal numbers (https://oeis.org/A051876)."""
 
-    sequence_name = "icositetragonal numbers"
+    SEQUENCE_NAME = "icositetragonal numbers"
 
     def __init__(self) -> None:
         super().__init__(number_of_sides=24)
@@ -364,10 +405,73 @@ class A051876(PolygonalNumbers):
 IcositetragonalNumbers = A051876
 
 
+class A053735(DigitSumSequence):
+    """Sum of digits of n written in base 3 (https://oeis.org/A053735)."""
+
+    SEQUENCE_NAME = "sequence A053735"
+
+    def __init__(self) -> None:
+        super().__init__(base=3)
+
+
+class A053737(DigitSumSequence):
+    """Sum of digits of n written in base 4 (https://oeis.org/A053737)."""
+
+    SEQUENCE_NAME = "sequence A053737"
+
+    def __init__(self) -> None:
+        super().__init__(base=4)
+
+
+class A053824(DigitSumSequence):
+    """Sum of digits of n written in base 5 (https://oeis.org/A053824)."""
+
+    SEQUENCE_NAME = "sequence A053824"
+
+    def __init__(self) -> None:
+        super().__init__(base=5)
+
+
+class A053827(DigitSumSequence):
+    """Sum of digits of n written in base 6 (https://oeis.org/A053827)."""
+
+    SEQUENCE_NAME = "sequence A053827"
+
+    def __init__(self) -> None:
+        super().__init__(base=6)
+
+
+class A053828(DigitSumSequence):
+    """Sum of digits of n written in base 7 (https://oeis.org/A053828)."""
+
+    SEQUENCE_NAME = "sequence A053828"
+
+    def __init__(self) -> None:
+        super().__init__(base=7)
+
+
+class A053829(DigitSumSequence):
+    """Sum of digits of n written in base 8 (https://oeis.org/A053829)."""
+
+    SEQUENCE_NAME = "sequence A053829"
+
+    def __init__(self) -> None:
+        super().__init__(base=8)
+
+
+class A053830(DigitSumSequence):
+    """Sum of digits of n written in base 9 (htpps://oeis.org/A053830)."""
+
+    SEQUENCE_NAME = "sequence A053830"
+
+    def __init__(self) -> None:
+        super().__init__(base=9)
+
+
 class A167149(PolygonalNumbers):
     """Myriagonal numbers (https://oeis.org/A167149)."""
 
-    sequence_name = "myriagonal numbers"
+    SEQUENCE_NAME = "myriagonal numbers"
 
     def __init__(self) -> None:
         super().__init__(number_of_sides=10_000)