+++ sequence/sequences/integer/recursive.py
@@ -12,7 +12,7 @@ from sequence.sequences.integer.recursive_generalised import (
 class A000032(LucasSequenceV):
     """Lucas numbers (https://oeis.org/A000032)."""
 
-    sequence_name = "Lucas numbers"
+    SEQUENCE_NAME = "Lucas numbers"
 
     def __init__(self) -> None:
         super().__init__(p=1, q=-1)
@@ -29,7 +29,7 @@ LucasNumbers = A000032
 class A000045(LucasSequenceU):
     """Fibonacci numbers (https://oeis.org/A000045)."""
 
-    sequence_name = "Fibonacci numbers"
+    SEQUENCE_NAME = "Fibonacci numbers"
 
     def __init__(self) -> None:
         super().__init__(p=1, q=-1)
@@ -42,7 +42,7 @@ FibonacciSequence = A000045
 class A000073(HighOrderFibonacciNumbers):
     """Tribonacci numbers (https://oeis.org/A000073)."""
 
-    sequence_name = "Tribonacci numbers"
+    SEQUENCE_NAME = "Tribonacci numbers"
 
     def __init__(self) -> None:
         super().__init__(order=3)
@@ -54,7 +54,7 @@ TribonacciNumbers = A000073
 class A000078(HighOrderFibonacciNumbers):
     """Tetranacci numbers (https://oeis.org/A000078)."""
 
-    sequence_name = "Tetranacci numbers"
+    SEQUENCE_NAME = "Tetranacci numbers"
 
     def __init__(self) -> None:
         super().__init__(order=4)
@@ -66,7 +66,7 @@ TetranacciNumbers = A000078
 class A000129(LucasSequenceU):
     """Pell numbers (https://oeis.org/A000129)."""
 
-    sequence_name = "Pell numbers"
+    SEQUENCE_NAME = "Pell numbers"
 
     def __init__(self) -> None:
         super().__init__(p=2, q=-1)
@@ -79,7 +79,7 @@ LambdaNumbers = A000129
 class A000931(AlmostMonotonicIncreasingMixin, Recursive):
     """Padovan numbers (https://oeis.org/A000931)"""
 
-    sequence_name = "Padovan numbers"
+    SEQUENCE_NAME = "Padovan numbers"
     offset: ClassVar[List[int]] = [1, 0, 0, 1, 0, 1]
 
     def __init__(self) -> None:
@@ -96,7 +96,7 @@ PadovanSequence = A000931
 class A001591(HighOrderFibonacciNumbers):
     """Pentanacci numbers (https://oeis.org/A001591)."""
 
-    sequence_name = "Pentanacci numbers"
+    SEQUENCE_NAME = "Pentanacci numbers"
 
     def __init__(self) -> None:
         super().__init__(order=5)
@@ -108,7 +108,7 @@ PentanacciNumbers = A001591
 class A001592(HighOrderFibonacciNumbers):
     """Hexanacci numbers (https://oeis.org/A001592)."""
 
-    sequence_name = "Hexanacci numbers"
+    SEQUENCE_NAME = "Hexanacci numbers"
 
     def __init__(self) -> None:
         super().__init__(order=6)
@@ -120,7 +120,7 @@ HexanacciNumbers = A001591
 class A001608(AlmostMonotonicIncreasingMixin, Recursive):
     """Perrin numbers (https://oeis.org/A001608)."""
 
-    sequence_name = "Perrin numbers"
+    SEQUENCE_NAME = "Perrin numbers"
     offset: ClassVar[List[int]] = [3, 0, 2, 3, 2, 5]
 
     def __init__(self) -> None:
@@ -136,7 +136,7 @@ PerrinNumbers = A001608
 class A002203(LucasSequenceV):
     """Companion Pell numbers (https://oeis.org/A002203)."""
 
-    sequence_name = "Companion Pell numbers"
+    SEQUENCE_NAME = "Companion Pell numbers"
 
     def __init__(self) -> None:
         super().__init__(p=2, q=-1)
@@ -149,7 +149,7 @@ PellLucasNumbers = A002203
 class A079262(HighOrderFibonacciNumbers):
     """Octanacci numbers (https://oeis.org/A079262)."""
 
-    sequence_name = "Octanacci numbers"
+    SEQUENCE_NAME = "Octanacci numbers"
 
     def __init__(self) -> None:
         super().__init__(order=8)
@@ -161,7 +161,7 @@ OctanacciNumbers = A079262
 class A104144(HighOrderFibonacciNumbers):
     """Enneanacci numbers (https://oeis.org/A104144)."""
 
-    sequence_name = "Enneanacci numbers"
+    SEQUENCE_NAME = "Enneanacci numbers"
 
     def __init__(self) -> None:
         super().__init__(order=9)
@@ -173,7 +173,7 @@ EnneanacciNumebrs = A104144
 class A122189(HighOrderFibonacciNumbers):
     """Heptanacci numbers (https://oeis.org/A122189)."""
 
-    sequence_name = "Heptanacci numbers"
+    SEQUENCE_NAME = "Heptanacci numbers"
 
     def __init__(self) -> None:
         super().__init__(order=7)
@@ -185,7 +185,7 @@ HeptanacciNumbers = A122189
 class A214733(LucasSequenceU):
     """Sequence A214733 (https://oeis.org/A214733)."""
 
-    sequence_name = "sequence A214733"
+    SEQUENCE_NAME = "sequence A214733"
 
     def __init__(self) -> None:
         super().__init__(p=-1, q=3)