+++ tests/tests_core/test_functions.py
@@ -1,0 +1,17 @@
+import pytest
+
+from sequence.core.utils.functions import digit_sum
+
+
+@pytest.mark.parametrize(
+    "number, base, expected_output",
+    [
+        (123, 10, 6),
+        (0, 10, 0),
+        (123, 2, 6),
+        (123, 4, 9),
+    ],
+)
+def test_digit_sum(number, base, expected_output) -> None:
+    if digit_sum(number=number, base=base) != expected_output:
+        raise ValueError(f"Expected: {expected_output}, got: {digit_sum(number=number, base=base)}")